MAJ 150608

+++ manage.py
@@ -16,13 +16,24 @@ def createadminuser():
     """
     Create Admin User in the database admin/password
     """
-    from appli import db,user_datastore
+    from appli import db,user_datastore,database
+    from appli.database import roles
+    r=roles.query.filter_by(id=1).first()
+    if r is None:
+        db.session.add(roles(id=1,name=database.AdministratorLabel))
+        db.session.commit()
+    r=roles.query.filter_by(id=2).first()
+    if r is None:
+        db.session.add(roles(id=2,name='Users Administrator'))
+        db.session.commit()
+
     u=user_datastore.find_user(email='admin')
     print(u)
-    user_datastore.delete_user(u)
-    user_datastore.create_user(email='admin', password=encrypt_password('altidev'))
-    user_datastore.create_role(name="admin",description="Application Administrator")
-    user_datastore.add_role_to_user('admin','admin')
+    if u is not None:
+        user_datastore.delete_user(u)
+        db.session.commit()
+    user_datastore.create_user(email='admin', password=encrypt_password('altidev'),name="Application Administrator")
+    user_datastore.add_role_to_user('admin','Application Administrator')
     db.session.commit()
 
 @manager.command
@@ -34,7 +45,13 @@ def dbcreate():
     from appli import db,user_datastore
     db.create_all()
 
-
+@manager.command
+def createsampledata():
+    from appli import db,database
+    r=database.Projects.query.filter_by(projid=1).first()
+    if r is None:
+        db.session.add(database.Projects(projid=1,title="Test Project 1"))
+        db.session.commit()
 
 
 if __name__ == "__main__":
